feat(refs DPLAN-12013): move component to own dir
